Place vertex types and instances in a mod
