+++ webrender/src/lib.rs
@@ -107,6 +107,7 @@ mod texture_allocator;
 mod texture_cache;
 mod tiling;
 mod util;
+mod vertex_types;
 
 mod shader_source {
     include!(concat!(env!("OUT_DIR"), "/shaders.rs"));
